# internal/api/handlers/security_session_handler.go
package handlers

import (
	"net/http"
	"time"

	"github.com/gin-gonic/gin"

	"thanawy-backend/internal/db"
	"thanawy-backend/internal/middleware"
	"thanawy-backend/internal/models"
)

// GetActiveSessions returns all active sessions
// @Summary Get active sessions
// @Description Get all active sessions for the current user or all users (admin)
// @Tags admin,security
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /api/admin/security/sessions [get]
func GetActiveSessions(c *gin.Context) {
	userID, _ := c.Get("userId")
	isAdmin := c.GetBool("is_admin")

	query := db.DB.Model(&models.UserSession{}).Where(statusQuery, "active")

	// Regular users can only see their own sessions
	if !isAdmin {
		query = query.Where(userIDQuery, userID)
	}

	var sessions []models.UserSession
	if err := query.Order("last_active_at DESC").Find(&sessions).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to fetch sessions"})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"data": gin.H{
			"sessions": sessions,
			"count":    len(sessions),
		},
	})
}

// RevokeSession revokes a specific session
// @Summary Revoke session
// @Description Revoke/end a specific session
// @Tags admin,security
// @Accept json
// @Produce json
// @Param id path string true "Session ID"
// @Success 200 {object} map[string]string
// @Router /api/admin/security/sessions/{id}/revoke [post]
func RevokeSession(c *gin.Context) {
	sessionID := c.Param("id")
	adminID, _ := c.Get("userId")

	var session models.UserSession
	if err := db.DB.First(&session, idQuery, sessionID).Error; err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Session not found"})
		return
	}

	session.Status = "revoked"
	now := time.Now()
	session.RevokedAt = &now
	adminIDStr := adminID.(string)
	session.RevokedBy = &adminIDStr

	if err := db.DB.Save(&session).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to revoke session"})
		return
	}

	middleware.LogCriticalOperation(c, "session_revoked", map[string]interface{}{
		"session_id": sessionID,
		"user_id":    session.UserID,
	})

	c.JSON(http.StatusOK, gin.H{"message": "Session revoked successfully"})
}

// RevokeOtherSessions revokes all sessions except current
// @Summary Revoke other sessions
// @Description Revoke all other sessions except the current one
// @Tags admin,security
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /api/admin/security/sessions/revoke-others [post]
func RevokeOtherSessions(c *gin.Context) {
	userID, _ := c.Get("userId")
	currentSessionID := c.GetString("session_id")

	result := db.DB.Model(&models.UserSession{}).
		Where("user_id = ? AND id != ? AND status = ?", userID, currentSessionID, "active").
		Updates(map[string]interface{}{
			"status":     "revoked",
			"revoked_at": time.Now(),
			"revoked_by": userID,
		})

	middleware.LogCriticalOperation(c, "sessions_revoked_others", map[string]interface{}{
		"revoked_count": result.RowsAffected,
	})

	c.JSON(http.StatusOK, gin.H{
		"message": "Other sessions revoked",
		"data": gin.H{
			"revokedCount": result.RowsAffected,
		},
	})
}

// RevokeUserSessions revokes all sessions for a specific user
// @Summary Revoke user sessions
// @Description Revoke all sessions for a specific user (admin only)
// @Tags admin,security
// @Accept json
// @Produce json
// @Param userId path string true "User ID"
// @Success 200 {object} map[string]interface{}
// @Router /api/admin/security/sessions/user/{userId}/revoke-all [post]
func RevokeUserSessions(c *gin.Context) {
	userID := c.Param("userId")
	adminID, _ := c.Get("userId")

	result := db.DB.Model(&models.UserSession{}).
		Where("user_id = ? AND status = ?", userID, "active").
		Updates(map[string]interface{}{
			"status":     "revoked",
			"revoked_at": time.Now(),
			"revoked_by": adminID,
		})

	middleware.LogCriticalOperation(c, "user_sessions_revoked", map[string]interface{}{
		"target_user":   userID,
		"revoked_count": result.RowsAffected,
	})

	c.JSON(http.StatusOK, gin.H{
		"message": "All user sessions revoked",
		"data": gin.H{
			"revokedCount": result.RowsAffected,
		},
	})
}

// GetSessionStats returns session statistics
// @Summary Get session statistics
// @Description Get statistics about user sessions
// @Tags admin,security
// @Accept json
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Router /api/admin/security/sessions/stats [get]
func GetSessionStats(c *gin.Context) {
	var stats struct {
		TotalActive   int64 `json:"totalActive"`
		TotalExpired  int64 `json:"totalExpired"`
		UniqueDevices int64 `json:"uniqueDevices"`
	}

	db.DB.Model(&models.UserSession{}).Where(statusQuery, "active").Count(&stats.TotalActive)
	db.DB.Model(&models.UserSession{}).Where(statusQuery, "expired").Count(&stats.TotalExpired)
	db.DB.Model(&models.UserSession{}).Where(statusQuery, "active").Select("COUNT(DISTINCT device_id)").Scan(&stats.UniqueDevices)

	c.JSON(http.StatusOK, gin.H{
		"data": stats,
	})
}